+++ Editor/Version/UtilityPanel.cs
@@ -6,39 +6,39 @@ namespace EtienneEditor
     public class UtilityPanel : EditorWindow
     {
 
-        private const string windowName = "Etienne Utility Panel";
+        private const string _WindowName = "Etienne Utility Panel";
 
-        private bool useDefaultLoader,
-            goBackToCurrentScene,
-            saveCurrentScene;
-        private int defaultSceneBuildIndex;
-        private string[] buildSceneNames;
+        private bool _UseDefaultLoader;
+        private bool _GoBackToCurrentScene;
+        private bool _SaveCurrentScene;
+        private int _DefaultSceneBuildIndex;
+        private string[] _BuildSceneNames;
 
-        [MenuItem("Tools/Etienne/" + windowName, priority = -100)]
+        [MenuItem("Tools/Etienne/" + _WindowName, priority = -100)]
         private static void ShowWindow()
         {
-            UtilityPanel window = GetWindow<UtilityPanel>(true, windowName);
-            window.minSize = new UnityEngine.Vector2(369, 400);
+            UtilityPanel window = GetWindow<UtilityPanel>(true, _WindowName);
+            window.minSize = new Vector2(369, 400);
             window.maxSize = window.minSize;
         }
 
         private void UpdateFields()
         {
-            useDefaultLoader = EditorPrefs.GetBool(EditorPrefsKeys.UseDefaultLoader, false);
-            defaultSceneBuildIndex = EditorPrefs.GetInt(EditorPrefsKeys.DefaultSceneBuildIndex, 0);
-            goBackToCurrentScene = EditorPrefs.GetBool(EditorPrefsKeys.GoBackToCurrentScene, true);
-            saveCurrentScene = EditorPrefs.GetBool(EditorPrefsKeys.AutoSaveCurrentScene, false);
+            _UseDefaultLoader = PrefsKeys.UseDefaultLoader;
+            _DefaultSceneBuildIndex = PrefsKeys.DefaultSceneBuildIndex;
+            _GoBackToCurrentScene = PrefsKeys.GoBackToCurrentScene;
+            _SaveCurrentScene = PrefsKeys.AutoSaveCurrentScene;
 
-            buildSceneNames = new string[EditorBuildSettings.scenes.Length];
-            for(int i = 0; i < buildSceneNames.Length; i++)
+            _BuildSceneNames = new string[EditorBuildSettings.scenes.Length];
+            for(int i = 0; i < _BuildSceneNames.Length; i++)
             {
                 string name = System.IO.Path.GetFileNameWithoutExtension(EditorBuildSettings.scenes[i].path);
-                buildSceneNames[i] = name;
+                _BuildSceneNames[i] = name;
             }
-            if(defaultSceneBuildIndex >= buildSceneNames.Length)
+            if(_DefaultSceneBuildIndex >= _BuildSceneNames.Length)
             {
-                defaultSceneBuildIndex = 0;
-                EditorPrefs.SetInt(EditorPrefsKeys.DefaultSceneBuildIndex, 0);
+                _DefaultSceneBuildIndex = 0;
+                PrefsKeys.DefaultSceneBuildIndex.SetValue(_DefaultSceneBuildIndex);
             }
         }
         private void OnEnable()
@@ -58,64 +58,87 @@ namespace EtienneEditor
 
         private void OnGUI()
         {
-            UpdatePackageGUI();
+            DrawVersion();
 
             EditorGUI.BeginChangeCheck();
-            useDefaultLoader = EditorGUILayout.BeginToggleGroup("Use default loader", useDefaultLoader);
-            if(EditorGUI.EndChangeCheck()) { EditorPrefs.SetBool(EditorPrefsKeys.UseDefaultLoader, useDefaultLoader); }
-
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("Default scene", EditorStyles.boldLabel);
-            UnityEngine.GUIStyle rightAlignedStyle = new UnityEngine.GUIStyle(EditorStyles.label) { alignment = UnityEngine.TextAnchor.MiddleRight, padding = new UnityEngine.RectOffset(0, 5, 0, 0) };
-            for(int i = 0; i < buildSceneNames.Length; i++)
+            _UseDefaultLoader = EditorGUILayout.BeginToggleGroup("Use default loader", _UseDefaultLoader);
+            if(EditorGUI.EndChangeCheck())
             {
-                EditorGUILayout.BeginHorizontal();
-                UnityEngine.Rect rect = EditorGUILayout.GetControlRect();
-
-                EditorGUI.LabelField(rect, new UnityEngine.GUIContent(i.ToString()), rightAlignedStyle);
-
-                EditorGUI.BeginChangeCheck();
-                EditorGUI.Toggle(rect, buildSceneNames[i], defaultSceneBuildIndex == i);
-                if(EditorGUI.EndChangeCheck()) { EditorPrefs.SetInt(EditorPrefsKeys.DefaultSceneBuildIndex, i); defaultSceneBuildIndex = i; }
-                EditorGUILayout.EndHorizontal();
+                PrefsKeys.UseDefaultLoader.SetValue(_UseDefaultLoader);
             }
+
+            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
+
+            DrawScenesBox();
+            DrawSettings();
+
             EditorGUILayout.EndVertical();
-
-            EditorGUI.BeginChangeCheck();
-            saveCurrentScene = EditorGUILayout.Toggle("Auto save current scene", saveCurrentScene);
-            if(EditorGUI.EndChangeCheck()) { EditorPrefs.SetBool(EditorPrefsKeys.AutoSaveCurrentScene, saveCurrentScene); }
-
-            EditorGUI.BeginChangeCheck();
-            goBackToCurrentScene = EditorGUILayout.Toggle("Go back to current scene", goBackToCurrentScene);
-            if(EditorGUI.EndChangeCheck()) { EditorPrefs.SetBool(EditorPrefsKeys.GoBackToCurrentScene, goBackToCurrentScene); }
 
             EditorGUILayout.EndToggleGroup();
-            EditorGUILayout.EndVertical();
-
         }
 
-        private static void UpdatePackageGUI()
+        private void DrawSettings()
+        {
+            EditorGUI.BeginChangeCheck();
+            _SaveCurrentScene = EditorGUILayout.Toggle("Auto save current scene", _SaveCurrentScene);
+            if(EditorGUI.EndChangeCheck())
+            {
+                PrefsKeys.AutoSaveCurrentScene.SetValue(_SaveCurrentScene);
+            }
+
+            EditorGUI.BeginChangeCheck();
+            _GoBackToCurrentScene = EditorGUILayout.Toggle("Go back to current scene", _GoBackToCurrentScene);
+            if(EditorGUI.EndChangeCheck())
+            {
+                PrefsKeys.GoBackToCurrentScene.SetValue(_GoBackToCurrentScene);
+            }
+        }
+
+        private void DrawScenesBox()
+        {
+            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
+
+            EditorGUILayout.LabelField("Default scene", EditorStyles.boldLabel);
+            for(int i = 0; i < _BuildSceneNames.Length; i++)
+            {
+                EditorGUILayout.BeginHorizontal();
+
+                Rect rect = EditorGUILayout.GetControlRect();
+
+                GUIStyle rightAlignedStyle = new GUIStyle(EditorStyles.label) { alignment = TextAnchor.MiddleRight, padding = new RectOffset(0, 5, 0, 0) };
+                EditorGUI.LabelField(rect, new GUIContent(i.ToString()), rightAlignedStyle);
+
+                EditorGUI.BeginChangeCheck();
+                EditorGUI.Toggle(rect, _BuildSceneNames[i], _DefaultSceneBuildIndex == i);
+                if(EditorGUI.EndChangeCheck())
+                {
+                    _DefaultSceneBuildIndex = i;
+                    PrefsKeys.DefaultSceneBuildIndex.SetValue(_DefaultSceneBuildIndex);
+                }
+
+                EditorGUILayout.EndHorizontal();
+            }
+
+            EditorGUILayout.EndVertical();
+        }
+
+        private static void DrawVersion()
         {
             GUIStyle style = new GUIStyle(GUI.skin.label)
             {
                 richText = true
             };
-            string currentVersion = VersionChecker.CurrentVersion;
-            string urlVersion = VersionChecker.UrlVersion;
-            bool hasCurrent = currentVersion != "0.0.0";
-            bool hasUrl = urlVersion != "0.0.0";
-            if(hasCurrent && hasUrl)
-            {
-                EditorGUILayout.BeginHorizontal();
-                string color = VersionChecker.IsUpToDate() ? "green" : "red";
-                GUIContent label = new GUIContent((hasCurrent ? "" : $"Current version: <color={color}>" + currentVersion + "</color>")
-                    + (!hasCurrent && !hasUrl ? ", " : "")
-                    + (hasUrl ? "" : "Newest version: " + urlVersion));
-                EditorGUILayout.LabelField(label, style);
-                EditorGUILayout.EndHorizontal();
-            }
-            if(UnityEngine.GUILayout.Button("Check Package")) VersionChecker.CheckVersion();
+
+            EditorGUILayout.BeginHorizontal();
+
+            string color = VersionChecker.IsUpToDate() ? "green" : "red";
+            GUIContent label =
+                new GUIContent($"Current version: <color={color}> {PrefsKeys.PackageCurrentVersion.GetValue()} </color>, Newest version: {PrefsKeys.PackageUrlVersion.GetValue()}");
+            EditorGUILayout.LabelField(label, style);
+
+            EditorGUILayout.EndHorizontal();
+
+            if(GUILayout.Button("Check Package")) VersionChecker.CheckVersion();
         }
     }
 }